r tar ut lenke til historikk som egen komponent
