+++ src/pages/index.tsx
@@ -5,6 +5,7 @@ import ManglerPersonEllerEnhet from '../components/feilmeldinger/mangler-person-
 import ArbeidssoekerstatusOversiktV2 from '../components/arbeidssoekerstatus-oversikt-v2';
 import HvaErNytt from '../components/hva-er-nytt';
 import DemoLabel from '../components/demo-label';
+import LenkeTilHistorikk from '../components/lenke-til-historikk';
 
 export default function Home() {
     return (
@@ -14,6 +15,7 @@ export default function Home() {
             <HvaErNytt />
             <ArbeidssoekerstatusOversiktV2 />
             <ManglerPersonEllerEnhet />
+            <LenkeTilHistorikk />
             <section className="flex flex-col items-center p-8">
                 <DemoPanel />
             </section>
